+++ src/stats.rs
@@ -8,7 +8,7 @@ pub fn degrees(r: f64) -> f64 {
 
 //https://rust-lang-nursery.github.io/rust-cookbook/science/mathematics/statistics.html
 pub fn mean(data: &[f32]) -> Option<f32> {
-    let sum = data.iter().sum::<f32>() as f32;
+    let sum = data.iter().sum::<f32>();
     let count = data.len();
 
     match count {
@@ -23,7 +23,7 @@ pub fn std_deviation(data: &[f32]) -> Option<f32> {
             let variance = data
                 .iter()
                 .map(|value| {
-                    let diff = data_mean - (*value as f32);
+                    let diff = data_mean - (*value);
 
                     diff * diff
                 })
@@ -43,7 +43,7 @@ pub fn z_score(pixel_value: f32, data: &[f32]) -> Option<f32> {
 
     match (data_mean, data_std_deviation) {
         (Some(mean), Some(std_deviation)) => {
-            let diff = data_value as f32 - mean;
+            let diff = data_value - mean;
             Some(diff / std_deviation)
         }
         _ => None,
